router: Cleanup build/bundle/sea scripts
