+++ packages/router/.gitignore
@@ -1,1 +1,3 @@
 /node-standalone
+/bundle.cjs
+/dist
